+++ ccramic/utils/roi_utils.py
@@ -32,3 +32,12 @@ def subset_mask_outline_using_cell_id_list(mask_outline, original_mask, cell_id_
     mask_outline[~mask_bool] = 0
     # converted = (mask_outline * 255).clip(0, 255).astype(np.uint8)
     return np.array(Image.fromarray(mask_outline.astype(np.float32)).convert('RGB')).astype(np.uint8)
+
+def override_roi_gallery_blend_list(currently_selected: list, saved_blend_dict: dict=None,
+                                    saved_blend:str=None):
+    """
+    Override the roi gallery blend list (channels) if a saved blend is used
+    """
+    if saved_blend_dict and saved_blend and saved_blend in saved_blend_dict:
+        return [i for i in saved_blend_dict[saved_blend]]
+    return currently_selected